feat: implement comprehensive SEO improvements
- Add robots.ts to guide crawlers (disallow /api, /settings, /login, /my-projects)
- Add dynamic sitemap.ts that fetches all projects from API
- Expand root metadata with OpenGraph, Twitter cards, canonical URL, icons
- Add JSON-LD structured data (WebSite + Organization schema)
- Switch fonts from external Google Fonts links to next/font/google (self-hosted, faster)
- Add generateMetadata to project detail page for per-project titles, descriptions, OG tags
- Split project detail into server wrapper (metadata) + client component (interactivity)
- Add per-page metadata via layout.tsx for /builders, /building, /cities, /projects
- Add custom 404 (not-found.tsx) and error boundary (error.tsx) pages
