+++ src/app/cities/layout.tsx
@@ -1,0 +1,18 @@
+import { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "Cities",
+  description:
+    "GrowthX builders by city — see which cities are shipping the most projects.",
+  alternates: {
+    canonical: "https://built.growthx.club/cities",
+  },
+};
+
+export default function CitiesLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return children;
+}
