+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
-/* eslint-disable @next/next/no-page-custom-font */
 import type { Metadata } from "next";
+import { DM_Sans, DM_Mono, Newsreader } from "next/font/google";
 import "./globals.css";
 import { LoginDialogProvider } from "@/context/LoginDialogContext";
 import { NavProvider } from "@/context/NavContext";
@@ -8,9 +8,75 @@ import LoginDialog from "@/components/LoginDialog";
 import AppNav from "@/components/AppNav";
 import CanvasFlip from "@/components/CanvasFlip";
 
+const dmSans = DM_Sans({
+  subsets: ["latin"],
+  variable: "--sans",
+  display: "swap",
+  weight: ["300", "400", "500", "600", "700"],
+});
+
+const dmMono = DM_Mono({
+  subsets: ["latin"],
+  variable: "--mono",
+  display: "swap",
+  weight: ["400", "500"],
+});
+
+const newsreader = Newsreader({
+  subsets: ["latin"],
+  variable: "--serif",
+  display: "swap",
+  weight: ["400", "500", "600"],
+  style: ["normal", "italic"],
+});
+
 export const metadata: Metadata = {
-  title: "Built at GrowthX",
-  description: "Discover what the GrowthX community is shipping",
+  metadataBase: new URL("https://built.growthx.club"),
+  title: {
+    default: "Built at GrowthX",
+    template: "%s — Built at GrowthX",
+  },
+  description:
+    "Discover what the GrowthX community is shipping — projects, builders, and city stats from India's top product community.",
+  openGraph: {
+    type: "website",
+    siteName: "Built at GrowthX",
+    title: "Built at GrowthX",
+    description:
+      "Discover what the GrowthX community is shipping — projects, builders, and city stats.",
+    images: ["/built-logo.svg"],
+  },
+  twitter: {
+    card: "summary",
+    title: "Built at GrowthX",
+    description:
+      "Discover what the GrowthX community is shipping — projects, builders, and city stats.",
+  },
+  icons: {
+    icon: "/built-logo.svg",
+    apple: "/built-logo.svg",
+  },
+  alternates: {
+    canonical: "https://built.growthx.club",
+  },
+};
+
+const jsonLd = {
+  "@context": "https://schema.org",
+  "@type": "WebSite",
+  name: "Built at GrowthX",
+  url: "https://built.growthx.club",
+  description:
+    "Discover what the GrowthX community is shipping — projects, builders, and city stats from India's top product community.",
+  publisher: {
+    "@type": "Organization",
+    name: "GrowthX",
+    url: "https://growthx.club",
+    logo: {
+      "@type": "ImageObject",
+      url: "https://built.growthx.club/built-logo.svg",
+    },
+  },
 };
 
 export default function RootLayout({
@@ -19,22 +85,20 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" suppressHydrationWarning>
+    <html
+      lang="en"
+      suppressHydrationWarning
+      className={`${dmSans.variable} ${dmMono.variable} ${newsreader.variable}`}
+    >
       <head>
+        <script
+          type="application/ld+json"
+          dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }}
+        />
         <script
           dangerouslySetInnerHTML={{
             __html: `(function(){try{var t=localStorage.getItem("bx-theme");if(t==="dark")document.documentElement.setAttribute("data-theme","dark")}catch(e){}requestAnimationFrame(function(){requestAnimationFrame(function(){document.documentElement.classList.add("ready")})})})()`,
           }}
-        />
-        <link rel="preconnect" href="https://fonts.googleapis.com" />
-        <link
-          rel="preconnect"
-          href="https://fonts.gstatic.com"
-          crossOrigin="anonymous"
-        />
-        <link
-          href="https://fonts.googleapis.com/css2?family=DM+Mono:wght@400;500&family=DM+Sans:ital,opsz,wght@0,9..40,300;0,9..40,400;0,9..40,500;0,9..40,600;0,9..40,700;1,9..40,400&family=Newsreader:ital,opsz,wght@0,6..72,400;0,6..72,500;0,6..72,600;1,6..72,400;1,6..72,500&display=swap"
-          rel="stylesheet"
         />
       </head>
       <body className="font-sans bg-bg text-text antialiased">
